restrict user who try to updates and deletes users

+++ backend/src/users/users.resolver.ts
@@ -6,6 +6,7 @@ import { CreateOneUserArgs } from 'src/generated/user/create-one-user.args';
 import { UpdateOneUserArgs } from 'src/generated/user/update-one-user.args';
 import { DeleteOneUserArgs } from 'src/generated/user/delete-one-user.args';
 import { User } from 'src/generated/user/user.model';
+import { CurrentUser } from 'src/users/decorators/currentUser.decorator';
 
 @Resolver((of: any) => User)
 export class UsersResolver {
@@ -28,17 +29,23 @@ export class UsersResolver {
     return this.usersService.createUser(data);
   }
 
-  //TODO リクエストを送信したユーザーの更新のみ許可する
   @Mutation((returns) => User) async updateUser(
+    @CurrentUser() currentUser: User,
     @Args() { data, where }: UpdateOneUserArgs,
   ): Promise<User> {
+    if (currentUser.id !== where.id)
+      throw new Error('You can only update your own user');
+
     return this.usersService.updateUser({ where, data });
   }
 
-  //TODO リクエストを送信したユーザーの削除のみ許可する
   @Mutation((returns) => User) async deleteUser(
+    @CurrentUser() currentUser: User,
     @Args() { where }: DeleteOneUserArgs,
   ): Promise<User> {
+    if (currentUser.id !== where.id)
+      throw new Error('You can only update your own user');
+
     return this.usersService.deleteUser(where);
   }
 }
